requirements.txt
pyirsdk==1.3.5
debugpy==1.6.7